backend/requirements.txt
fastapi
uvicorn
python-dotenv
passlib[argon2]
python-jose
motor
pymongo
groq
